[Update] move redux steps to QueryBuilderHandler
